requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9

faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4

pypdf==4.2.0

openai==1.42.0

celery==5.4.0
redis==5.0.8

slowapi==0.1.9
